requirements/origin.txt
# --- Libs ---

# requests
httpie
# pillow
# pygraphviz
# pyparsing
# pydot
# pydotplus

# PyYAML
# uritemplate
# Markdown
Pygments

# jupyter

# --- Science ---
# scipy
# numpy
# matplotlib
# pandas


# --- Framework ---

# pyqt5


## -- Web --

# Flask
# flask_sqlalchemy
# Flask-Testing

django
djangorestframework
# django-debug-toolbar
# django-extensions
# django-filter
# django-guardian
# dj-database-url

### - Config -
# configparser

### - Static file -
# whitenoise


# --- Server ---

# gunicorn


# --- Database ---

# psycopg2
# mysql-connector-python		# do not use mysql-connector
# influxdb-client[ciso]
# influxdb-client
# influxdb




